# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Interior Design Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LeadInputProjectType = typeof LeadInputProjectType[keyof typeof LeadInputProjectType];


export const LeadInputProjectType = {
  new_construction: 'new_construction',
  renovation: 'renovation',
  redecoration: 'redecoration',
} as const;

export type LeadInputBudget = typeof LeadInputBudget[keyof typeof LeadInputBudget];


export const LeadInputBudget = {
  under_10k: 'under_10k',
  '10k_25k': '10k_25k',
  '25k_50k': '25k_50k',
  '50k_100k': '50k_100k',
  over_100k: 'over_100k',
} as const;

export type LeadInputTimeline = typeof LeadInputTimeline[keyof typeof LeadInputTimeline];


export const LeadInputTimeline = {
  immediately: 'immediately',
  '1_3_months': '1_3_months',
  '3_6_months': '3_6_months',
  '6_12_months': '6_12_months',
  flexible: 'flexible',
} as const;

export type LeadInputStyle = typeof LeadInputStyle[keyof typeof LeadInputStyle];


export const LeadInputStyle = {
  modern: 'modern',
  contemporary: 'contemporary',
  traditional: 'traditional',
  scandinavian: 'scandinavian',
  industrial: 'industrial',
  bohemian: 'bohemian',
  minimalist: 'minimalist',
  mid_century: 'mid_century',
} as const;

export interface LeadInput {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  projectType: LeadInputProjectType;
  roomTypes: string[];
  budget: LeadInputBudget;
  timeline: LeadInputTimeline;
  style: LeadInputStyle;
  /** @nullable */
  squareFootage: number | null;
  description: string;
  /** @nullable */
  hearAboutUs?: string | null;
}

export interface Lead {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  projectType: string;
  roomTypes: string[];
  budget: string;
  timeline: string;
  style: string;
  /** @nullable */
  squareFootage: number | null;
  description: string;
  /** @nullable */
  hearAboutUs?: string | null;
  createdAt: string;
}

export interface StatCount {
  label: string;
  count: number;
}

export interface LeadStats {
  totalLeads: number;
  byStyle: StatCount[];
  byBudget: StatCount[];
  byProjectType: StatCount[];
}

export interface ErrorResponse {
  error: string;
}