# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Interior Design Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Lead {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  projectType: string;
  roomTypes: string[];
  budget: string;
  timeline: string;
  style: string;
  /** @nullable */
  squareFootage: number | null;
  description: string;
  /** @nullable */
  hearAboutUs?: string | null;
  createdAt: string;
}