# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Interior Design Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { LeadInputBudget } from './leadInputBudget';
import type { LeadInputProjectType } from './leadInputProjectType';
import type { LeadInputStyle } from './leadInputStyle';
import type { LeadInputTimeline } from './leadInputTimeline';

export interface LeadInput {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  projectType: LeadInputProjectType;
  roomTypes: string[];
  budget: LeadInputBudget;
  timeline: LeadInputTimeline;
  style: LeadInputStyle;
  /** @nullable */
  squareFootage: number | null;
  description: string;
  /** @nullable */
  hearAboutUs?: string | null;
}