# lib/api-zod/src/generated/types/leadStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Interior Design Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { StatCount } from './statCount';

export interface LeadStats {
  totalLeads: number;
  byStyle: StatCount[];
  byBudget: StatCount[];
  byProjectType: StatCount[];
}