# lib/api-zod/src/generated/types/statCount.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Interior Design Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface StatCount {
  label: string;
  count: number;
}